project/requirements.txt
Django==1.8.3
djangorestframework==3.4.0
django-rest-swagger==0.3.10
django-supervisor==0.3.2
#python-social-auth
#django-tts
#django-cleanup==0.4.2
gunicorn==19.6.0
pillow==2.3.0

